--- a/Engine/DataFeeds/UpdateData.cs
+++ b/Engine/DataFeeds/UpdateData.cs
@@ -27,6 +27,13 @@ namespace QuantConnect.Lean.Engine.DataFeeds
     public class UpdateData<T>
     {
         /// <summary>
+        /// Flag indicating whether <see cref="Data"/> contains any fill forward bar or not
+        /// </summary>
+        /// <remarks>This is useful for performance, it allows consumers to skip re enumerating the entire data
+        /// list to filter any fill forward data</remarks>
+        public readonly bool? ContainsFillForwardData;
+
+        /// <summary>
         /// The target, such as a security or subscription data config
         /// </summary>
         public readonly T Target;
@@ -55,12 +62,14 @@ namespace QuantConnect.Lean.Engine.DataFeeds
         /// <param name="data">The update data</param>
         /// <param name="isInternalConfig">True if this update data corresponds to an internal subscription
         /// such as currency or security benchmark</param>
-        public UpdateData(T target, Type dataType, IReadOnlyList<BaseData> data, bool isInternalConfig)
+        /// <param name="containsFillForwardData">True if this update data contains fill forward bars</param>
+        public UpdateData(T target, Type dataType, IReadOnlyList<BaseData> data, bool isInternalConfig, bool? containsFillForwardData = null)
         {
             Target = target;
             Data = data;
             DataType = dataType;
             IsInternalConfig = isInternalConfig;
+            ContainsFillForwardData = containsFillForwardData;
         }
     }
 }