--- a/Queues/JobQueue.cs
+++ b/Queues/JobQueue.cs
@@ -40,7 +40,6 @@ namespace QuantConnect.Queues
         private static readonly int UserId = Config.GetInt("job-user-id", 0);
         private static readonly int ProjectId = Config.GetInt("job-project-id", 0);
         private readonly string AlgorithmTypeName = Config.Get("algorithm-type-name");
-        private readonly string AlgorithmPathPython = Config.Get("algorithm-path-python", "../../../Algorithm.Python/");
         private readonly Language Language = (Language)Enum.Parse(typeof(Language), Config.Get("algorithm-language"));
 
         /// <summary>
@@ -177,7 +176,7 @@ namespace QuantConnect.Queues
                 if (!File.Exists(pythonSource))
                 {
                     // Copies file to execution location
-                    foreach (var file in new DirectoryInfo(AlgorithmPathPython).GetFiles("*.py"))
+                    foreach (var file in new DirectoryInfo(Path.GetDirectoryName(AlgorithmLocation)).GetFiles("*.py"))
                     {
                         file.CopyTo(file.FullName.Replace(file.DirectoryName, Environment.CurrentDirectory), true);
                     }